* Components/ArticleTable/ArticleTablePlugin.h         * Components/ArticleView/ArticleTextViewPlugin.m         * NS(U)Integer transitions
git-svn-id: svn+ssh://svn.savannah.nongnu.org/gap/trunk@2260 d25b4930-3292-e311-b074-0010c6de5da8

+++ user-apps/Grr/Components/ArticleTable/ArticleTablePlugin.h
@@ -58,12 +58,12 @@
 
 // ---------------- NSTableView data source ----------------------
 
-- (int) numberOfRowsInTableView: (NSTableView *)aTableView;
+- (NSInteger) numberOfRowsInTableView: (NSTableView *)aTableView;
 
 
 - (id)           tableView: (NSTableView *)aTableView
  objectValueForTableColumn: (NSTableColumn *)aTableColumn
-                       row: (int)rowIndex;
+                       row: (NSInteger)rowIndex;
 
 // ------------------- NSTableView delegate ------------------------
